Add disclaimer gate before main website
