+++ src/app/home/page.tsx
@@ -1,0 +1,5 @@
+import { MainWebsiteContent } from "@/components/marketing/main-website-content";
+
+export default function HomePage() {
+    return <MainWebsiteContent />;
+}
